refactor: extract child run termination tool
